feat(types): better typings of find/get/findAll
This commit allows to have a better type inference of `find`, `get` and `findAll` methods.
For example, `wrapper.find('input').element` is now inferred as a `HTMLInputElement | undefined`. It also works with SVG element: `wrapper.get('line').element` is inferred as a `SVGLineElement`. For ambiguous selectors, we fallback to the generic element type: `wrapper.get('#other').element` is inferred as `Element`.