feat: getComponent
This introduces a `getComponent` which returns a `VueWrapper` or throws.
It offers two benefits:
- a symetric API for components/elements with `findComponent/findAllComponents/getComponent`/`find/findAll/get`
- a better experience for TS users (as `findComponent` returns a union type, you can't chain certain things until you disambiguated the type manually).

I also refactored the signatures of `find` to offer proper overloads (the current version breaks tests compilation in a real app, i hope this fixes it). I added TSD tests to ensure the inference is working properly.