chore: remove ts-expect-error in tsd tests
It turns out you need to remove `ts-expect-error` in tsd tests with more recent versions of tsd.
It makes sense: tsd uses TS diagnostics to know if there is an error or not.
As older tsd versions used a TS version that did not understand `ts-expect-error` there was no issue.
Now that tsd does use a TS version supporting this comment, the tests started to fail.