fix: emits can be an object
The fix introduced in #521 only account for emits define as an array, but it can also be define as an object.
The latest rc.5 release breaks on components that use an object to define emits with:

    TypeError: emits.includes is not a function

This commit fixes the unit test to have both cases check, and the relevant code to properly handle the object case.