fix(mount): shallowMount with Vue v3.2.32
v3.2.32 changed how the root component is used in `createApp` https://github.com/vuejs/core/commit/d121a9bc7e7af59adb2d2803954cfeee95b35270

The root component is now copied before usage.

This broke how we stub components, as we used to add the root component to the list of component to not stub.
As the component is now no longer the same, all tests using shallowMount are failing.

The fix is to register the copy of the component created by Vue in `createApp` (which is available as `app._component`)