chore: use --bundleConfigAsCjs to load rollup config
Ideally, we'd like to use a `mjs` config, but that necessitates an import type assertion for the import of `package.json`,
and this feature is not supported by node v14.
So we'll be able to use a `rollup.config.mjs` config only when we'll stop supporting node v14.
See https://github.com/vuejs/test-utils/pull/1815