docs: add section about test runners upgrade to migration guide
Context: https://github.com/vuejs/vue-test-utils/issues/1975#issuecomment-13595554
As discussed, there a couple of issues that the community faced while upgrading
to vue@3 @vue/test-utils@^2 but that are not directly related with them.

In a effort to make migrations more smooth, a new section  was added
to the migration docs that lists some common problems with underlying
test runners, that might happen simply because of the fact that
users took the opportunity to upgrade their test stacks dependecies
as well.

It starts with a few issues with jest, one reported in the
legacy test-utils repo and other one that i've personally faced.

https://github.com/vuejs/vue-test-utils/issues/1975