fix: TS workaround for Vue v3.2.47
A change of types in core (https://github.com/vuejs/core/pull/2818) introduced a TS compilation error when bumping to Vue v3.2.47.
I couldn't figure out a proper way to fix, but using `EmitOptions` instead of `E` (which extends `EmitOptions`...) fixes the issue.
It may be good enough to allow to update to the latest Vue releases, and I think it should have minimal impact on our users (or none at all).