fix(stubs): avoid warning on normalized props with Vue v3.4.22
Fixes #2411

We can "trick" the warning introduced in Vue v3.4.22 (See https://github.com/vuejs/core/commit/7ccd453dd004076cad49ec9f56cd5fe97b7b6ed8)
by using an array of strings instead of a string (as the new check allows functions and arrays, but not strings).
This does not affect the rendering of the stub, and still displays `[Function]`, so it should not impact the existing tests.